CORE-1645: WIP fixing literals in version properties in gradle...
