--- applications/examples/test-cpk/build.gradle
@@ -1,17 +1,0 @@
-plugins {
-    id 'net.corda.plugins.cordapp-cpk'
-}
-
-cordapp {
-    targetPlatformVersion platformVersion as Integer
-    minimumPlatformVersion 1
-    contract {
-        name "test-cpk"
-        versionId 1
-        vendor "R3"
-    }
-}
-
-dependencies {
-    cordaProvided "org.jetbrains.kotlin:kotlin-osgi-bundle:$kotlinVersion"
-}
