--- applications/examples/test-cpk/src/main/kotlin/net/corda/sample/testcpk/TestCPK.kt
@@ -1,16 +1,0 @@
-package net.corda.sample.testcpk
-
-/**
- * Dummy CordApp used in the `:applications:examples:test-sandbox`.
- */
-class TestCPK: Runnable {
-
-    init {
-        println("net.corda.sample.testcpk.TestCPK INIT.")
-    }
-
-
-    override fun run() {
-        println("net.corda.sample.testcpk.TestCPK RUN.")
-    }
-}
