--- applications/examples/test-sandbox/detekt-baseline.xml
@@ -1,4 +1,0 @@
-<SmellBaseline>
-    <ManuallySuppressedIssues></ManuallySuppressedIssues>
-    <CurrentIssues></CurrentIssues>
-</SmellBaseline>
